requirements.txt
matplotlib==3.8.2
numpy==1.24.3
pandas==1.5.3
Pillow==10.2.0
streamlit==1.29.0
torch==2.1.1
opencv-python==4.5.3.56
git+https://github.com/ultralytics/yolov5.git@v5.0
